Realigned zitilib/examples/tiny and zitilib/examples/diamondback with the characterization and mattermozt models. (#25)
